requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.6
djangorestframework==3.15.2
idna==3.10
requests==2.32.3
sqlparse==0.5.3
urllib3==2.3.0
django-redis
